requirements.txt
pytest
requests
jsonschema
pytest-html
setuptools
python-dotenv